requirements.txt
# CDN Server Selection Project - Python Requirements
# Python 3.8+ required
# Install with: pip install -r requirements.txt

# ============================================
# Core Data Science Libraries
# ============================================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0

# ============================================
# Machine Learning Frameworks
# ============================================
scikit-learn>=1.3.0,<2.0.0
xgboost>=2.0.0,<3.0.0
lightgbm>=4.0.0,<5.0.0

# Optional: Deep Learning (uncomment if needed)
# tensorflow>=2.13.0,<3.0.0
# torch>=2.0.0,<3.0.0
# torchvision>=0.15.0,<1.0.0

# ============================================
# Data Visualization
# ============================================
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.14.0,<6.0.0

# ============================================
# Network Measurement Tools
# ============================================
ripe.atlas.cousteau>=1.4.2
ripe.atlas.tools>=3.0.0
requests>=2.31.0,<3.0.0
dnspython>=2.3.0,<3.0.0

# ============================================
# BigQuery & M-Lab Data Access
# ============================================
google-cloud-bigquery>=3.38.0,<4.0.0
pandas-gbq>=0.30.0,<1.0.0

# ============================================
# Jupyter & Interactive Development
# ============================================
jupyter>=1.0.0
notebook>=6.5.0,<8.0.0
ipython>=8.12.0,<9.0.0
ipywidgets>=8.0.0,<9.0.0

# ============================================
# Data Processing & Utilities
# ============================================
tqdm>=4.65.0,<5.0.0
pyarrow>=12.0.0,<14.0.0  # Fast parquet reading
openpyxl>=3.1.0,<4.0.0   # Excel support
joblib>=1.3.0,<2.0.0     # Parallel processing

# ============================================
# Statistical Analysis
# ============================================
statsmodels>=0.14.0,<1.0.0
pingouin>=0.5.3,<1.0.0   # Advanced stats

# ============================================
# Geospatial & Network Analysis
# ============================================
geopandas>=0.13.0,<1.0.0
folium>=0.14.0,<1.0.0
networkx>=3.1,<4.0.0

# ============================================
# Configuration & Environment
# ============================================
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0.0
toml>=0.10.2

# ============================================
# Testing & Quality
# ============================================
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
flake8>=6.0.0,<7.0.0
black>=23.0.0,<24.0.0

# ============================================
# API & Web Scraping (Optional)
# ============================================
beautifulsoup4>=4.12.0,<5.0.0
selenium>=4.10.0,<5.0.0

# ============================================
# Performance Monitoring
# ============================================
psutil>=5.9.0,<6.0.0
memory_profiler>=0.61.0,<1.0.0

# ============================================
# Date/Time Handling
# ============================================
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3
